FI-765 fix: directly end `e2ed` on termination signals
fix: add more temporary logs on termination signals
