+++ bin/dockerEntrypoint.sh
@@ -16,6 +16,7 @@ onExit() {
         echo "PID $PID is running"
         kill -USR1 $PID
 
+        echo "dockerEntrypoint will sleep ${E2ED_TIMEOUT_FOR_GRACEFUL_SHUTDOWN_IN_SECONDS:-16} for seconds"
         sleep "${E2ED_TIMEOUT_FOR_GRACEFUL_SHUTDOWN_IN_SECONDS:-16}"
     fi
